backend/pyproject.toml
[project]
name = "wordbird"
version = "0.3.3"
description = "Voice dictation daemon using NVIDIA Parakeet on Apple Silicon"
readme = "../README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mlx-audio",
    "sounddevice",
    "numpy",
    "pyobjc-core",
    "pyobjc-framework-Cocoa",
    "pyobjc-framework-Quartz",
    "python-frontmatter",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "tomli; python_version < '3.11'",
]

[dependency-groups]
dev = ["pytest"]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wordbird"]

[project.scripts]
wordbird = "wordbird.cli:main"
wordbird-server = "wordbird.cli:server_main"
wordbird-daemon = "wordbird.cli:daemon_main"
